% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mme.R
\name{calculate_mme}
\alias{calculate_mme}
\title{Calculate morphine milligram equivalents (MME)}
\usage{
calculate_mme(x, ...)
}
\arguments{
\item{x}{(\code{list}, \code{data.frame}, or \code{tbl_df})\cr
Object with input data; either a \code{list} of medications or long format
data - \code{data.frame} or tibble.}

\item{...}{Additional arguments passed to methods}
}
\description{
\itemize{
\item \code{\link[=calculate_mme.list]{calculate_mme.list()}}
\item \code{\link[=calculate_mme.data.frame]{calculate_mme.data.frame()}}
\item \code{\link[=calculate_mme.tbl_df]{calculate_mme.tbl_df()}}
}
}
\keyword{internal}
