% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_rate_limit_helpers.R
\name{check_unique_pat}
\alias{check_unique_pat}
\title{Check if API rate limit warning is needed}
\usage{
check_unique_pat(n_patients, use_api, rate_limit = 50)
}
\arguments{
\item{n_patients}{Number of unique patients}

\item{use_api}{Whether API is being used}

\item{rate_limit}{The API rate limit (default: 50)}
}
\value{
Invisible NULL, called for side effects (warnings)
}
\description{
Check if API rate limit warning is needed
}
\keyword{internal}
