% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_med_list.R
\name{get_med_list}
\alias{get_med_list}
\title{Retrieve full opioid medication list}
\usage{
get_med_list()
}
\value{
A \code{data.frame} with full list of opioid medication names (\code{med_name}) that are compatible with the MME calculator along with their conversion factors (\code{cf}).
}
\description{
Retrieve full opioid medication list
}
\examples{
get_med_list()

}
\seealso{
\code{\link[=search_meds]{search_meds()}}
}
