% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmequiv-package.R
\docType{package}
\name{mmequiv-package}
\alias{mmequiv}
\alias{mmequiv-package}
\title{mmequiv: Calculate Standardized Morphine Milligram Equivalent Doses in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Uses standardized methods to calculate morphine milligram equivalents (MME) for opioid dose comparison. The package includes API wrapper functions to call NIH HEAL MME Online Calulator and local functions that replicate the API's calculations on the user's local machine from the comfort of R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kennethataylor.github.io/mmequiv/}
}

}
\author{
\strong{Maintainer}: Kenneth A. Taylor \email{kenneth.taylor.dpt@gmail.com} (\href{https://orcid.org/0000-0002-3205-9280}{ORCID}) [copyright holder]

}
\keyword{internal}
