% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mmequiv_is_transient}
\alias{mmequiv_is_transient}
\title{Check if MME Calculator API response indicates transient error}
\usage{
mmequiv_is_transient(resp)
}
\arguments{
\item{resp}{A httr2 response object.}
}
\value{
A logical value indicating whether there was a rate limit error (429) and
no API calls remaining.
}
\description{
Check if MME Calculator API response indicates transient error
}
\keyword{internal}
