% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mmequiv_req_retry}
\alias{mmequiv_req_retry}
\title{Automatically retry a request on failure with \code{mmequiv} package details}
\usage{
mmequiv_req_retry(req, ...)
}
\arguments{
\item{req}{A httr2 request object.}
}
\value{
A modified HTTP (\code{httr2}) request.
}
\description{
Automatically retry a request on failure with \code{mmequiv} package details
}
\keyword{internal}
