% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_meds.R
\name{search_meds}
\alias{search_meds}
\title{Search opioid medication list}
\usage{
search_meds(med_name = NULL)
}
\arguments{
\item{med_name}{A single string specifying the medication name to search for.}
}
\value{
A \code{data.frame} containing medications matching the \code{med_name} argument and
their associated conversion factor(s) (\code{cf}).
}
\description{
Search opioid medication list
}
\examples{
search_meds("oxy")

}
\seealso{
\code{\link[=get_med_list]{get_med_list()}}
}
