% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmequiv-package.R
\docType{package}
\name{mmequiv-package}
\alias{mmequiv}
\alias{mmequiv-package}
\title{mmequiv: Calculate Standardized Morphine Milligram Equivalent Doses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate morphine milligram equivalents (MME) for opioid dose comparison using standardized methods. Includes API wrapper functions to call 'NIH HEAL MME Online Calculator' and functions that replicate API calculations on the user's local machine from the comfort of 'R'. Creation of the 'NIH HEAL MME Online Calculator' and the MME calculations implemented in this package are described in Adams MCB, Sward KA, Perkins ML, Hurley RW (2025) \doi{10.1097/j.pain.0000000000003529}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kennethataylor.github.io/mmequiv/}
  \item \url{https://github.com/KennethATaylor/mmequiv}
  \item Report bugs at \url{https://github.com/KennethATaylor/mmequiv/issues}
}

}
\author{
\strong{Maintainer}: Kenneth A. Taylor \email{kenneth.taylor.dpt@gmail.com} (\href{https://orcid.org/0000-0002-3205-9280}{ORCID}) [copyright holder]

}
\keyword{internal}
