% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmequiv-package.R
\docType{package}
\name{mmequiv-package}
\alias{mmequiv}
\alias{mmequiv-package}
\title{mmequiv: R Interface to the NIH HEAL Morphine Milligram Equivalent API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Wraps the NIH HEAL Morphine Milligram Equivalent (MME) Online Calculator API, which standardizes research methods for opioid dose comparison. The functions in this API wrapper allow users to access and implement all of the NIH HEAL MME Online Calulator functions from the comfort of R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kennethataylor.github.io/mmequiv/}
}

}
\author{
\strong{Maintainer}: Kenneth A. Taylor \email{kenneth.taylor.dpt@gmail.com} (\href{https://orcid.org/0000-0002-3205-9280}{ORCID})

}
\keyword{internal}
